refactor(svelte2tsx): `$$render()` ラッパー生成を create_render_function.rs へ
`///<reference types="svelte" />` ヘッダの前置と `function $$render() { … }` の
開き部分の生成を `createRenderFunction.ts` 相当のファイルへ切り出す。
`$$props` / `$$restProps` / `$$slots` 宣言テキストの組み立ても同じファイルへ移した。

純粋な移動でロジック変更はなく、出力はバイト単位で不変。
