+++ crates/rsvelte_core/src/svelte2tsx/mod.rs
@@ -1,4 +1,5 @@
 pub mod add_component_export;
+pub mod create_render_function;
 pub mod helpers;
 pub mod interfaces;
 #[allow(
